refactoring works for dafny run
